refactor: findProductById 수정 및 적용
