+++ app/src/main/java/nextstep/shoppingcart/ui/shoppingcart/component/ShoppingCartItemHeader.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import nextstep.shoppingcart.R.string.shopping_cart_item_header_close_button
-import nextstep.shoppingcart.data.dummyProducts
+import nextstep.shoppingcart.data.Products.dummyProducts
 import nextstep.shoppingcart.ui.shoppinglist.model.Product
 import nextstep.shoppingcart.ui.theme.RobotoBold
