+++ app/src/main/java/nextstep/shoppingcart/ui/shoppinglist/ShoppingListScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.R.string.shopping_list_title
-import nextstep.shoppingcart.data.dummyProducts
+import nextstep.shoppingcart.data.Products.dummyProducts
 import nextstep.shoppingcart.ui.shoppinglist.component.ShoppingListLazyVerticalGrid
 import nextstep.shoppingcart.ui.shoppinglist.component.ShoppingListTopBar
 import nextstep.shoppingcart.ui.shoppinglist.model.Product
